Move Monitor components under monitor feature
