+++ src/monitor/components/useMonitorMapLayerSync.ts
@@ -1,7 +1,7 @@
 import type { StormReport } from '../../types/stormReports';
-import type { NwsAlertFeatureCollection } from '../../monitor/nwsAlerts';
-import type { MonitorMapView } from '../../monitor/types';
-import type { WmsLayerConfig } from '../../monitor/wms';
+import type { NwsAlertFeatureCollection } from '../nwsAlerts';
+import type { MonitorMapView } from '../types';
+import type { WmsLayerConfig } from '../wms';
 import type { SerializedMonitorOutlookFeature } from './monitorMapFeatureSync';
 import type { useMonitorMapRefs } from './monitorMapRefs';
 import { useMonitorMapOverlaySync } from './useMonitorMapOverlaySync';
@@ -24,6 +24,7 @@ interface UseMonitorMapLayerSyncArgs {
   onClearSelectedAlert: () => void;
 }
 
+/** Coordinates WMS and vector overlay synchronization for the monitor map. */
 export const useMonitorMapLayerSync = (args: UseMonitorMapLayerSyncArgs) => {
   useMonitorMapWmsSync({
     darkMode: args.darkMode,
